Redo the dataset creation
This is the changes that we tried to do to be able to compute the confusion matrix. Since I got confused and computed it before training, now the problem is how to train this model and then compute the confusion matrix. For this, I have to understand how to make the data pipeline work so that the original stay the same and I still get to have the labels.